+++ Source/Encoding/x265Enc.vb
@@ -41,9 +41,9 @@ Public Class x265Enc
         End Get
     End Property
 
-    Overrides ReadOnly Property ResizingStatus As String
+    Overrides ReadOnly Property IsUnequalResizingAllowed As Boolean
         Get
-            Return If(String.IsNullOrWhiteSpace(Params.GetStringParam(Params.DolbyVisionRpu.Switch)?.Value), "", "Resizing would interfere with the Dolby Vision metadata. Disable the 'Resize' filter or remove the Dolby Vision RPU file.")
+            Return String.IsNullOrWhiteSpace(Params.GetStringParam(Params.DolbyVisionRpu.Switch)?.Value)
         End Get
     End Property